[Fix] 배달 구조 수정
